add mapper support
