+++ pkg/memory/memory_test.go
@@ -9,7 +9,7 @@ import (
 
 func TestMemoryImmediate(t *testing.T) {
 	t.Parallel()
-	m := New(nil, nil, nil, nil)
+	m := New(nil, nil, nil, nil, nil)
 
 	i := new(uint8)
 	m.WriteMemoryAddressModes(1, i)
@@ -21,7 +21,7 @@ func TestMemoryImmediate(t *testing.T) {
 
 func TestMemoryAbsoluteInt(t *testing.T) {
 	t.Parallel()
-	m := New(nil, nil, nil, nil)
+	m := New(nil, nil, nil, nil, nil)
 
 	m.WriteMemoryAddressModes(1, 2)
 	assert.Equal(t, 1, m.ReadMemory(2))
@@ -34,7 +34,7 @@ func TestMemoryAbsoluteInt(t *testing.T) {
 
 func TestMemoryAbsoluteIndirect(t *testing.T) {
 	t.Parallel()
-	m := New(nil, nil, nil, nil)
+	m := New(nil, nil, nil, nil, nil)
 	x := new(uint8)
 	y := new(uint8)
 	m.LinkRegisters(x, y, x, y)
@@ -53,14 +53,14 @@ func TestMemoryAbsoluteIndirect(t *testing.T) {
 }
 
 func TestReadMemory16(t *testing.T) {
-	m := New(nil, nil, nil, nil)
+	m := New(nil, nil, nil, nil, nil)
 	m.WriteMemory(0, 1)
 	m.WriteMemory(1, 2)
 	assert.Equal(t, 0x201, m.ReadMemory16(0))
 }
 
 func TestWriteMemory16(t *testing.T) {
-	m := New(nil, nil, nil, nil)
+	m := New(nil, nil, nil, nil, nil)
 	m.WriteMemory16(0, 0x201)
 	assert.Equal(t, 0x201, m.ReadMemory16(0))
 }
