+++ pkg/system/system.go
@@ -5,6 +5,7 @@ import (
 	"github.com/retroenv/nesgo/pkg/cartridge"
 	"github.com/retroenv/nesgo/pkg/controller"
 	"github.com/retroenv/nesgo/pkg/cpu"
+	"github.com/retroenv/nesgo/pkg/mapper"
 	"github.com/retroenv/nesgo/pkg/memory"
 	"github.com/retroenv/nesgo/pkg/ppu"
 )
@@ -24,14 +25,19 @@ type System struct {
 }
 
 // New creates a new NES system.
-func New(cartridge *cartridge.Cartridge) *System {
+func New(cart *cartridge.Cartridge) *System {
+	mapp, err := mapper.New(cart)
+	if err != nil {
+		panic(err)
+	}
+
 	sys := &System{
-		PPU:         ppu.New(memory.NewRAM(0x2000)),
+		PPU:         ppu.New(memory.NewRAM(0x2000), mapp),
 		Controller1: controller.New(),
 		Controller2: controller.New(),
 	}
 
-	sys.Memory = memory.New(cartridge, sys.PPU, sys.Controller1, sys.Controller2)
+	sys.Memory = memory.New(cart, sys.PPU, sys.Controller1, sys.Controller2, mapp)
 	sys.CPU = cpu.New(sys.Memory, &sys.IrqHandler)
 	return sys
 }
